requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
deezer-python==6.1.1
defusedxml==0.7.1
ffmpeg-python==0.2.0
flatbencode==0.2.1
future==0.18.3
idna==3.4
iso8601==2.1.0
m3u8==3.6.0
musicbrainzngs==0.7.1
mutagen==1.47.0
Pillow==10.1.0
protobuf==3.15.8
pycryptodomex==3.19.0
python-dotenv==1.0.0
python-qbittorrent==0.4.3
requests==2.31.0
six==1.16.0
torf==4.2.2
tqdm==4.66.1
urllib3==2.0.7
